Refactor: Update imports, service paths, and add null checks
