+++ frontend/src/app/pages/department-management/department-management.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit, Inject, PLATFORM_ID } from '@angular/core';
 import { CommonModule, isPlatformBrowser } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
-import { NotificationService } from './notification.service';
+import { NotificationService } from '../../services/notification.service';
 
 interface Department {
   id: number;
