+++ frontend/src/app/pages/dict-management/dict-item-management.component.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { Router, ActivatedRoute } from '@angular/router';
-import { NotificationService } from './notification.service';
+import { NotificationService } from '../../services/notification.service';
 
 interface DictItem {
   id?: number;
